Issue #SB-17524 feat: displaying the offline screen for youtube contnet when internet gose off
