+++ src/app/client/projects/desktop/src/app/modules/content-consumption/components/content-player/content-player.component.ts
@@ -4,7 +4,7 @@ import { Component, AfterViewInit, ViewChild, ElementRef, Input, Output, EventEm
 import * as _ from 'lodash-es';
 import { PlayerConfig } from '@sunbird/shared';
 import { Router } from '@angular/router';
-import { ToasterService, ResourceService } from '@sunbird/shared';
+import { ToasterService, ResourceService , OfflineCardService} from '@sunbird/shared';
 const OFFLINE_ARTIFACT_MIME_TYPES = ['application/epub', 'video/webm', 'video/mp4', 'application/pdf'];
 import { Subject } from 'rxjs';
 import { ConnectionService, ContentManagerService } from '@sunbird/offline';
@@ -33,7 +33,7 @@ export class ContentPlayerComponent implements AfterViewInit, OnChanges, OnInit,
   @Input() isContentPresent = true;
   @Input() objectRollUp: {} = {};
   isConnected: any;
-  @Input() youTubeContentStatus = false;
+  youTubeContentStatus: any;
   public unsubscribe$ = new Subject<void>();
   CONSTANT = {
     ACCESSEVENT: 'renderer:question:submitscore'
@@ -48,6 +48,7 @@ export class ContentPlayerComponent implements AfterViewInit, OnChanges, OnInit,
     public resourceService: ResourceService, public navigationHelperService: NavigationHelperService,
     private connectionService: ConnectionService,
     private contentManagerService: ContentManagerService,
+    private offlineCardService: OfflineCardService,
     public playerService: PublicPlayerService) {
     this.buildNumber = (<HTMLInputElement>document.getElementById('buildNumber'))
       ? (<HTMLInputElement>document.getElementById('buildNumber')).value : '1.0';
@@ -77,6 +78,17 @@ export class ContentPlayerComponent implements AfterViewInit, OnChanges, OnInit,
       this.objectRollUp = _.get(this.playerConfig, 'context.objectRollup') || {};
       this.loadPlayer();
     }
+    this.handleYoutubeContent(this.contentData);
+  }
+  handleYoutubeContent(data) {
+    this.connectionService.monitor().pipe(takeUntil(this.unsubscribe$)).subscribe(isConnected => {
+      if (!isConnected && this.offlineCardService.isYoutubeContent(data)) {
+        this.youTubeContentStatus = true;
+      } else {
+        this.youTubeContentStatus = false;
+        this.loadPlayer();
+      }
+        });
   }
   ngOnInit() {
     this.contentManagerService.deletedContent.pipe(takeUntil(this.unsubscribe$)).subscribe((data) => {
