+++ src/app/client/projects/desktop/src/app/modules/content-consumption/components/contentplayer-page/contentplayer-page.component.html
@@ -11,7 +11,7 @@
 </div>
 
 <div *ngIf="tocPage" class="toc-content" [appTelemetryImpression]="telemetryImpression">
-    <app-content-player [playerConfig]="playerConfig" [contentData]="contentDetails" [isContentPresent]="isContentPresent" [youTubeContentStatus]= "youTubeContentStatus" [isContentDeleted]="isContentDeleted"></app-content-player>
+    <app-content-player [playerConfig]="playerConfig" [contentData]="contentDetails" [isContentPresent]="isContentPresent"  [isContentDeleted]="isContentDeleted"></app-content-player>
     <app-content-metadata [contentData]="contentDetails"></app-content-metadata>
 </div>
 <div *ngIf="!tocPage" class="sb-single-resource" [appTelemetryImpression]="telemetryImpression">
