+++ src/app/client/projects/desktop/src/app/modules/content-consumption/components/toc-page/toc-page.component.html
@@ -2,7 +2,7 @@
 <div class="ui container" [appTelemetryImpression]="telemetryImpression" *ngIf="collectionData">
   <div class="sb-g">
     <div class="sb-g-col-xs-12 sb-g-col-md-7 sb-g-col-lg-8">
-      <app-contentplayer-page [tocPage]="true" [contentDetails]="activeContent" [isContentPresent]="isContentPresent" [youTubeContentStatus]= "youTubeContentStatus" [objectRollUp]="objectRollUp"></app-contentplayer-page>
+      <app-contentplayer-page [tocPage]="true" [contentDetails]="activeContent" [isContentPresent]="isContentPresent"  [objectRollUp]="objectRollUp"></app-contentplayer-page>
     </div>
     <div class="sb-g-col-xs-12 sb-g-col-md-5 sb-g-col-lg-4">
       <sb-library-filters [list]="mimeTypeFilters" [layout]="'round-with-icon'"
